create total osx install script and update all

+++ install/install.sh
@@ -7,8 +7,16 @@ link_dir=$dotfiles/link
 # detect os
 OS="$(uname -s)"
 
+function iHeader(){ echo "\n\033[1m$@\033[0m\n"; }
+function iStep()  { echo "\033[1;33m➜\033[0m $@"; }
+function iGood()  { echo "\033[1;33m➜\033[0m $@"; }
+function iBad()   { echo "\033[1;31m✖\033[0m $@"; }
+
+iHeader "Starting bootstrap install @Gilad"
+
 # Ubuntu-only stuff
-if [[ "$OS" =~ Linux ]]
+if [[ "$OS" =~ Linux ]]; then
+    iHeader "Running Linux Setup"
     # specify the sudoers filename
     sudoers_file=sudoers-dotfiles
     # sudoers file location
@@ -17,60 +25,42 @@ if [[ "$OS" =~ Linux ]]
     sudoers_dest=/etc/sudoers.d/$sudoers_file
     # copy sudoers to sudoers.d
     if [[ ! -e "$sudoers_dest" || "$sudoers_dest" -ot "$sudoers_src" ]]; then
-        visudo -cf "$sudoers_src" >/dev/null && {
-        sudo cp "$sudoers_src" "$sudoers_dest" &&
-            sudo chmod 0440 "$sudoers_dest"
-    } >/dev/null 2>&1 && echo "File $sudoers_dest updated." || echo "Error updating $sudoers_dest file."
-fi
+        iStep "Setting up sudo account for easier installing"
+        visudo -cf "$sudoers_src" >/dev/null &&
+            sudo cp "$sudoers_src" "$sudoers_dest" &&
+            sudo chmod 0440 "$sudoers_dest" &&
+            echo "File $sudoers_dest updated." || echo "Error updating $sudoers_dest file."
+    fi
 fi
 
 # OSX-only stuff.
-if [[ "$OS" =~ ^Darwin ]]
+if [[ "$OS" =~ ^Darwin ]]; then
+    iHeader "Running OSX setup"
+
     # Install Homebrew.
     if [[ ! "$(type -p brew)" ]]; then
+        iStep "Installing Homebrew"
         # install homebrew
         ruby -e "$(curl -fsSL https://raw.github.com/Homebrew/homebrew/go/install)"
     fi
-
-    if [[ "$(type -p brew)" ]]; then
-        brew doctor
-        brew update
-
-        packages=(
-        git
-        git-extras
-        htop-osx
-        hub
-        id3tool
-        lesspipe
-        macvim
-        man2html
-        maven
-        mongo
-        nmap
-        nodejs
-        python
-        ruby
-        sl
-        tmux
-        tree
-        )
-        for package in $packages; do
-            brew install $package
-        done
-    fi
 fi
 
-#TODO handle already created files
-
 # create symlinks
-for filename in $link_dir/{.,_}*; do
-    echo "Creating symlink for $filename"
-    ln -sf $filename ~/
+iHeader "Creating symlinks"
+for filename in $link_dir/{.,_}[!.]*; do
+    if [[ ! -e "$filename" ]]; then
+        iGood "Symlink created: $filename"
+        ln -sf $filename ~/
+    else
+        iBad "Symlink skipped, file exists: $filename"
+    fi
 done
 
 # make vim bundles dir
+iHeader "Handling vim install"
 if [[ ! -d "$vim_bundles" ]]; then
-    echo "Creating vim bundles directory at $vim_bundles"
+    iStep "Creating vim bundles directory at $vim_bundles"
     mkdir -p "$vim_bundles"
 fi
+
+unset iHeader iStep iGood iBad
